change method of excluding padding in the encoder
Rewritten the way encoder_exclude_padding functions.
This now works only for TimestepMLP and MeanTimestepMLP, but is much more computationally efficient.
Using it for these two encoders is now recommended, particularly for MeanTimestepMLP (since this weights all timesteps the same, so the encoding in the beginning is heavily influenced by the padding).