MAJOR BUGFIX: Training latent GMM
The indentation in the reconstruction trainer was wrong. This meant that when training a GMM, we only ever trained the first component.

Training models with num_classes = 1 (i.e. a single latent Gaussian) was not affected.

The bug was introduced with commit "refactor encoder" (10.02.2022, 0ef48071a5c1e6c7613e82792a4d104a72f0a917)