add option to decode full episode groups
We now allow resetting an episode to the same task and storing these consecutive episodes with an identifier in the replay buffer.
When the decoder_window is specified as [-Inf, Inf], we now decode all of these consecutive episodes from the same z.
This way, we can collect task info over multiple episodes for understanding the task, but the actual task encoding is still based on a small time window.